Testing with small data
jupyter can't test the whole astronomy dataset because its too large. I've sorted the smileyface csv and am testing it that way first. The issue may be with the get neighbors function. The expandClusters and disjointset header do work with the code because it is a CPU function and doesn't directly work with CUDA.